LayoutTests/js/dom/modules/import-from-loaded-classic.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test import from loaded classic.');

// Module will be executed asynchronously.
window.jsTestIsAsync = true;

function loaded()
{
    shouldBeTrue(`window.imported`);
    finishJSTest();
}
</script>
<script>
</script>
<script src="../../../resources/js-test-post.js"></script>
<script src="./resources/import-from-loaded-classic.js"></script>
</body>
</html>
