Safari-611.1.10/LayoutTests/js/dom/modules/import-from-loaded-classic-expected.txt

Test import from loaded classic.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


import from loaded classic
import from loaded classic finish
PASS window.imported is true
PASS successfullyParsed is true

TEST COMPLETE
